Refactor stores
