--- src/common/types/Hash.ts
@@ -1,3 +1,0 @@
-export default interface Hash<T> {
-  [key: string]: T;
-}
