--- src/common/types/Profile.ts
@@ -1,3 +1,0 @@
-export default interface Profile {
-  id: string;
-}
