+++ src/components/NotificationManager/index.ts
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
-import { State } from "@/common/store/rootReducer";
-import { selectNotifications } from "@/common/store/notifications";
+import { State } from "@/stores/rootReducer";
+import { selectNotifications } from "@/stores/notifications";
 
 import NotificationManager from "./NotificationManager";
