+++ src/features/game/components/FullScreenButton.tsx
@@ -4,7 +4,7 @@ import FullscreenIcon from "@material-ui/icons/Fullscreen";
 import FullscreenExitIcon from "@material-ui/icons/FullscreenExit";
 
 import { createNotification } from "@/game/engine/notification";
-import { Severity } from "@/common/store/notifications";
+import { Severity } from "@/stores/notifications";
 import useWindowEvent from "@/hooks/useWindowEvent";
 import {
   isFullScreen,
