+++ src/features/game/components/StatusPanel.tsx
@@ -14,7 +14,7 @@ import elapsedGameTime from "@/game/utils/elapsedGameTime";
 import { StrokeStyles } from "@/game/enums/StrokeStyle";
 import logo from "@/assets/images/logo.svg";
 import { ProxyStoreConsumer } from "@/providers/ProxyStoreProvider";
-import { selectSettings, toggleSetting } from "@/common/store/settings";
+import { selectSettings, toggleSetting } from "@/stores/settings";
 import { useGripService } from "@/game/xstate/services";
 import { GripStrengthString } from "@/game/xstate/machines/gripMachine";
 import useForceUpdate from "@/hooks/useForceUpdate";
