+++ src/features/handy/components/ConnectHandy/ConnectHandy.tsx
@@ -19,7 +19,7 @@ import ContactlessIcon from "@material-ui/icons/Contactless";
 import { MenuItem } from "@/components/Templates";
 import useToggle from "@/hooks/useToggle";
 import { createNotification } from "@/game/engine/notification";
-import { Severity } from "@/common/store/notifications";
+import { Severity } from "@/stores/notifications";
 import useStickyState from "@/hooks/useStickyState";
 import useForceUpdate from "@/hooks/useForceUpdate";
