+++ src/features/search/components/CreatedGamesTable.tsx
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
-import { State } from "@/common/store/rootReducer";
-import { selectProfile } from "@/common/store/currentUser";
+import { State } from "@/stores/rootReducer";
+import { selectProfile } from "@/stores/currentUser";
 
 import GamesTable from "./GamesTable";
