+++ src/game/engine/notification.ts
@@ -2,12 +2,12 @@
  * Transitioning to new notification system, this file has been turned
  * into an adapter
  */
-import store from "@/common/store";
+import store from "@/stores";
 import {
   createNotification as createNotification2,
   dismissNotification as dismissNotification2,
   dismissAllNotifications as dismissAllNotifications2,
-} from "@/common/store/notifications";
+} from "@/stores/notifications";
 
 export const createNotification = (
   ...args: Parameters<typeof createNotification2>
