+++ src/test/test-utils.tsx
@@ -11,7 +11,7 @@ import { MemoryRouter as Router } from "react-router-dom";
 import Auth0TestProvider, {
   Auth,
 } from "@/providers/AuthProvider/Auth0TestProvider";
-import store from "@/common/store";
+import store from "@/stores";
 
 type Options = RenderOptions & {
   route?: string;
