+++ src/types/Hash.ts
@@ -1,0 +1,3 @@
+export type Hash<T> = {
+  [key: string]: T;
+};
