+++ src/types/Profile.ts
@@ -1,0 +1,3 @@
+export type Profile = {
+  id: string;
+};
